requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Django==4.1.5
django-embed-video==1.4.8
python-dotenv==0.20.0
django-channels==0.7.0
google-api-python-client==2.71.0
django-utils-six==2.0
channels==4.0.0
schedule==1.1.0
django-ratelimit==4.0.0